test(deploy: server/db/models/user.spec + /api/users.spec, component/auth-form + copyright + sign-up)passing all tests. copyright is a component and is imported in auth-form and sign-up.
