+++ server/db/models/user.spec.js
@@ -15,6 +15,8 @@ describe('User model', () => {
 
       beforeEach(async () => {
         cody = await User.create({
+          firstName: 'Cody',
+          lastName: 'the Pug',
           email: 'cody@puppybook.com',
           password: 'bones'
         })
